+++ vault.tf
@@ -1,5 +1,5 @@
 resource "vault_generic_secret" "default" {
-  path = "${var.vault_secret_path}/postgresql/passwd"
+  path      = "${var.vault_secret_path}/postgresql/passwd"
   data_json = <<EOT
 {
   "postgres": "${google_sql_user.postgres.password}",